HTBHF-485 Move determining the next allowed step and is path allowed into the state machine, plus stop the session being cleared when getting to the confirm page. Also make sure once at the COMPLETED state you can only see the confirm page.
