+++ src/web/routes/application/middleware/handle-path-request.js
@@ -1,18 +1,20 @@
-const { equals } = require('ramda')
 const { CHECK_URL, CONFIRM_URL } = require('../common/constants')
+const { stateMachine, actions } = require('../common/state-machine')
+
+const { IS_PATH_ALLOWED, GET_NEXT_ALLOWED_PATH } = actions
 
 const getPathsInSequence = (steps) => [...steps.map(step => step.path), CHECK_URL, CONFIRM_URL]
-
-const isPathAllowed = (sequence, allowed, path) =>
-  sequence.findIndex(equals(path)) <= sequence.findIndex(equals(allowed))
 
 const middleware = (pathsInSequence) => (req, res, next) => {
   if (!req.session.nextAllowedStep) {
     req.session.nextAllowedStep = pathsInSequence[0]
   }
 
-  if (!isPathAllowed(pathsInSequence, req.session.nextAllowedStep, req.path)) {
-    return res.redirect(req.session.nextAllowedStep)
+  const isPathAllowed = stateMachine.dispatch(IS_PATH_ALLOWED, req, pathsInSequence, req.session.nextAllowedStep, req.path)
+  const nextAllowedPath = stateMachine.dispatch(GET_NEXT_ALLOWED_PATH, req, req.session.nextAllowedStep)
+
+  if (!isPathAllowed) {
+    return res.redirect(nextAllowedPath)
   }
 
   next()
@@ -22,6 +24,5 @@ const handleRequestForPath = (steps) => middleware(getPathsInSequence(steps))
 
 module.exports = {
   getPathsInSequence,
-  isPathAllowed,
   handleRequestForPath
 }
