+++ src/web/routes/application/middleware/handle-path-request.test.js
@@ -1,34 +1,15 @@
 const test = require('tape')
 const sinon = require('sinon')
 const { CHECK_URL, CONFIRM_URL } = require('../common/constants')
-const { getPathsInSequence, isPathAllowed, handleRequestForPath } = require('./handle-path-request')
+const { getPathsInSequence, handleRequestForPath } = require('./handle-path-request')
 
 const steps = [{ path: '/first', next: '/second' }, { path: '/second' }]
-const paths = ['/first', '/second', '/third', '/fourth']
 
 test('getPathsInSequence() returns the correct sequence of paths', (t) => {
   const expected = ['/first', '/second', CHECK_URL, CONFIRM_URL]
   const result = getPathsInSequence(steps)
 
   t.deepEqual(result, expected, 'returns the correct sequence of paths')
-  t.end()
-})
-
-test('isPathAllowed() should return true if path is before allowed in sequence', (t) => {
-  const result = isPathAllowed(paths, '/third', '/second')
-  t.equal(result, true, 'returns true if path is before allowed in sequence')
-  t.end()
-})
-
-test('isPathAllowed() should return true if path matches allowed', (t) => {
-  const result = isPathAllowed(paths, '/third', '/third')
-  t.equal(result, true, 'returns true if path matches allowed')
-  t.end()
-})
-
-test('isPathAllowed() should return false if path is after allowed in sequence', (t) => {
-  const result = isPathAllowed(paths, '/third', '/fourth')
-  t.equal(result, false, 'returns false if path is after allowed in sequence')
   t.end()
 })
